update: scheme load func (lualine)
The original theme loading function relied totally on using files the
the lua/themes folder. Lualine provides a lot of themes out of the box
that are left unused by the sheme loading libary because of this
structure. This commit aims to fix this issue and load default lualine
themes without referencing files if non exist.